Fix JSONACTION_RETURN_SUCCESS definition on pg16-
On postgres 15 and below, pgsptokentype is defined to void rather than an
actual enum, so a lot of functions don't return anything.  But an empty
JSONACTION_RETURN_SUCCESS define doesn't exit the function as it would with
pg16+ definition, which leads to bogus behavior.  To fix, simply define
JSONACTION_RETURN_SUCCESS to "return" on pg 15 and below.