Add PostgreSQL 17 compatibility (#44)
Postgres 17 changed some of the lexer token numbering, so introduce a new
pgsp_token_types.h file to force the declaration of the needed token per major
version and avoid using wrong values.

The JsonLexContext structure changed again and added extra ability to free part
of the struct, which leads to wrong parsing results if not setup correctly.
Fortunately upstream makeJsonLexContextCstringLen() has been changed to handle
stack-allocated structs so we can now rely on this function from now on to
avoid misbehavior in case of further changes to JsonLexContext.

blk_read_time and blk_write_time BufferUsage fields have been renamed to
include a "shared" prefix, so handle it.

Finally, enable test on pg17 in the CI.

Co-authored-by: Devrim Gunduz <devrim@gunduz.org>